refactor(userspace-dp): extract helpers/planning module
Continue the #6234 decomposition: move the binding/queue planning domain
out of server/helpers.rs into helpers/planning.rs. This is the second of
four cohesive submodules.

Per server/README.md the settle predicates
(wait_for_binding_settle/bindings_settled), the canonical same-plan hash
key (snapshot_binding_plan_key and its JSON-canonicalizing helpers), and
the RX-queue / binding replanner (replan_queues,
replan_bindings_from_candidates, effective_rx_queues, plan_key_rx_queues,
include_userspace_binding_interface) form one correctness unit. The
design-review amendment on the issue explicitly required keeping hash
ownership and layout ownership together — separating them invites the
same-plan / stale-layout regressions fixed by #2915/#2916/#3007/#3175 —
so binding_plan and queue_plan land as a single planning.rs, not two.

Pure code-motion: bodies are byte-for-byte identical (extracted verbatim
from the pre-split source). Explicit per-module imports (crate::protocol
binding/snapshot types, crate::server::ServerState, sha2, chrono, std
collections/io/thread/time) replace the crate-root glob. The two
cross-module deps route through the helpers facade
(super::{refresh_status, should_run_afxdp}). mod.rs re-exports
planning::* so replan_queues, snapshot_binding_plan_key,
same_plan_apply_needs_binding_reconcile, and the rx-queue test overrides
keep their existing server::helpers::<name> call-site paths. The
now-unused sha2/std::io imports are dropped from mod.rs.

Validation: `cargo build` on userspace-dp is green with no new warnings
in helpers/.

Refs #6234
